Move superclass, clean up imports
